moves Ordering related unit tests to new OrderingByPositionTest class
